changes (BREAKING): rename :feature:movies root package for avoid conflicts with main app module.
