+++ features/movies/src/main/java/com/christopher_elias/features/movies/presentation/model/MovieUi.kt
@@ -1,0 +1,28 @@
+package com.christopher_elias.features.movies.presentation.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+/*
+ * Created by Christopher Elias on 26/04/2021
+ * christopher.mike.96@gmail.com
+ *
+ * Loop Ideas
+ * Lima, Peru.
+ */
+
+@Parcelize
+data class MovieUi(
+    val id: Int,
+    val isAdultOnly: Boolean,
+    val popularity: Double,
+    val voteAverage: Double,
+    val voteCount: Int,
+    val image: String,
+    val backdropImage: String,
+    val title: String,
+    val overview: String,
+    val releaseDate: String,
+    val originalTitle: String,
+    val originalLanguage: String
+) : Parcelable
